pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "kenforever",email = "ken@kenforever.cc"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "siwe (>=4.4.0,<5.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins.dotenv]
location = ".env.dev"
